add tags to offline query
