+++ serializers.go
@@ -7,13 +7,13 @@ import (
 	"time"
 )
 
-func (p OnlineQueryParams) serialize() onlineQueryRequestSerialized {
+func (p OnlineQueryParams) serialize() internal.OnlineQueryRequestSerialized {
 	context := internal.OnlineQueryContext{
 		Environment: internal.StringOrNil(p.EnvironmentId),
 		Tags:        p.Tags,
 	}
 
-	body := onlineQueryRequestSerialized{
+	body := internal.OnlineQueryRequestSerialized{
 		Inputs:           p.inputs,
 		Outputs:          p.outputs,
 		Context:          context,
@@ -157,7 +157,7 @@ func (c *ErrorCodeCategory) UnmarshalJSON(data []byte) error {
 }
 
 func (p OfflineQueryParams) MarshalJSON() ([]byte, error) {
-	queryInput := offlineQueryInputSerialized{}
+	queryInput := internal.OfflineQueryInputSerialized{}
 	globalInputTimes := make([]any, 0)
 
 	for fqn, tsFeatureValues := range p.inputs {
@@ -185,7 +185,7 @@ func (p OfflineQueryParams) MarshalJSON() ([]byte, error) {
 		requiredOutput = make([]string, 0)
 	}
 
-	serializedObj := offlineQueryRequestSerialized{
+	serializedObj := internal.OfflineQueryRequestSerialized{
 		Input:             queryInput,
 		Output:            output,
 		RequiredOutput:    requiredOutput,
@@ -193,6 +193,7 @@ func (p OfflineQueryParams) MarshalJSON() ([]byte, error) {
 		Branch:            internal.StringOrNil(p.Branch),
 		MaxSamples:        p.MaxSamples,
 		DestinationFormat: "PARQUET",
+		Tags:              p.Tags,
 	}
 
 	return json.Marshal(serializedObj)
